requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
docx2pdf>=0.1.8
fpdf2>=2.7.4
pywin32>=306; sys_platform == 'win32'